[move-cli] run meta tests in a separate process
If the meta test and other CLI test cases run in different threads but
in the same process, then their environment variables, including
`env::current_dir()`, are going to be shared. This is not good,
especially with the coverage tracking (`MOVE_VM_TRACE`) env var.

In fact, this might be the culprit of some flackiness in the tests, as
observed in #8917 and #8846 (although they are very rare and very hard
to reproduce locally). The reason can be some form of race condition
around the `MOVE_VM_TRACE` flag. One possible procedure is:

- meta test, test A, test B starts in different threads
- meta test set `MOVE_VM_TRACE=env::current_dir()/build/trace`
  but `env::current_dir()` returns `A/`
- test B invokes Move VM, Move VM sees the `MOVE_VM_TRACE` is set,
  and tries to create the trace file in `A/build`
- but `A/build` is already deleted, leading to a `unwrap()` error.

This is my guess on why in #8917, test `publish_then_check` tries to
access the `build` directory of `explain_user_tx_script_abort`, which is
a different and unrelated test.

Ideally, we should really run thest tests all in different processes, as
they are designed this way. Fortunately, most of the CLI tests run in
different processes already (via `Command::new()`), having the meta test
runs in a separate process will break the race condition on the
`MOVE_VM_TRACE` flag and hopefully solve the flakiness.