[move-prover] embed the modular verification logic in prover driver
The modular verification logic can be reached by specifying `--modular`
in the move-prover command line.

If this `--modular` flag is set, the verification is splitted into two
parts:
- whole package verification
- modular verification per input Move file.

Let's use the following case as an example:
`mvp --modular --dependency stdlib/ my-sources/`
And suppose `my-sources` has two move files, A.move and B.move.

It is equivalent to invoking a series of verification:

- step 1: package verification
  which is to verify A.move and B.move together (i.e., both are target modules)

- step 2: modular verificaiton, which is essentially similar to
  - `mvp --dependency stdlib/ --dependency my-sources/ my-sources/A.move`
  - `mvp --dependency stdlib/ --dependency my-sources/ my-sources/B.move`

This commit also adds modular verification to existing verification
tests (as part of unit tests). But these verification tests are marked
with `#[ignore]` so that they won't be invoked in CI by default. These
additional tests can be invoked with `cargo test -- --ignored`.